refactor codegen for tf
